stdlib: fix cpu utilization/cycles intermediate truncation
some of the stdlib routines are doing intermediate rounding which
causes less accurate final results to be produced. to minimize the
impact of this, we make two changes:
- pull "/ 1000" outside of SUM() aggregations so we aggregate first
 and then do the truncating divide.
- when "/ 1000" is in the numerator and denominator, remove it from
 both.

This does not fix all inaccuracies. as you can see in diff tests
there are still some cases where the inequality
min_freq <= avg_freq <= max_freq does not hold, which is still unexpected.