tracebox: improve default behaviour of tracebox on non-Android (#2747)
On all OSes today, tracebox defaults to using an abstract socket rather
than a system socket. This makes sense on Android as it's always
expected that the system daemons are running and we don't want to step
on them.

However, on Linux, it's very rare that someone has another daemon
running in the background. Instead, it's almost always the case that
someone is doing a "one-shot" trace collection and in that case we will
very much want to use system sockets to make sure that track_event, gpu
render stages (e.g. mesa), other data sources etc correctly connect ot
this instance of Perfetto.

The fact this doesn't work trips people up all the time: the mesa folk
are still using the tmux script because they though that something was
broken with tracebox: they didn't realise that --system-sockets was
necessary to correctly collect mesa traces.