traceconv: add support for new "bundling" mode (#2751)
The idea of this mode is that it's an all-in-one location to "bundle" a
trace file to make it shareable, work outside the machine the trace was
collected on etc. Today, all we're doing is symbolizing Linux traces but
in the future, we'll extend this to Android symbolization, deobfuscation
etc with a bunch of heuristics.