tp: AnalyzeStructuredQuery gets columns of the query (#2765)
We need to know the resulting columns of the query for Explore Page but
we don't want to have to require user to specify them.

To do it we need to execute preambles and modules, but it is not a
particularily bad behavior. We can split it into multiple functions if
it starts to be an issue.