ui: Remove `globals`; Move `isInternalUser` flag to AppImpl
We currently still rely on the global `globals` object for only a small
handful of things. `App` and `AppImpl` have largely taken over the role
of what `globals` used to offer, so in order to make things consistent
this has been merged into there and the `globals.ts` deleted to avoid
future code from depending on it.

The scripts `is_internal_user.js` and friends still write to `globals`
and this is difficult to change as these are third party scripts not
version controlled with the UI's codebase, so we set up a
`window.globals` proxy object that proxies any writes to `AppImpl`.
Crucially, this globals object is not exposed to TypeScript, so tsc
will complain if we start using it accidentially.

Change-Id: I780b8b7fdba815f1519ddaee3fa59b1f22fdd08b