exp: Fix filters on aggregation mode (#2785)
Filters on AggregationNode should be applied to the result of the
aggregation, not to the source columns like in the situation of all
other nodes.

The rest of the code is a trivial refactor of moving the responsibility
of where the filteres are applied from node_explorer to the nodes.