ui: Fix headless track selection bug (#2811)
When selecting *over* a headless track (users are unaware of this
because headless tracks are invisible), the entire group is selected as
this is the behavior when selecting over a non-headless track.

This patch fixes the issue by omitting headless tracks from the rendered
track list so the selection logic never even sees them.

Fixes: https://buganizer.corp.google.com/issues/442408089