ui: Add track settings infrastructure & bulk track settings edit menu (#2859)
Currently tracks own their settings completely, which is clean
encapsulation but makes bulk operations and global defaults impossible.

This patch adds infrastructure for 'track settings' where each track can
advertise a list of settings that it supports, including how to render
the menu items in so that the user can modify those settings, and also
getters and settings to get and modify the setting values externally.

This patch is built with bulk setting modification in mind, as each
settings has a descriptor associated with it, and multiple tracks can
point to the same descriptor. When multiple tracks are selected, the
list of common settings supported by all selected tracks is generated
(those that point to the same descriptor) are rendered in the bulk track
dropdown menu in the header panel of the timeline (see screenshot).

<img width="516" height="482" alt="image"
src="https://github.com/user-attachments/assets/230667d8-f0db-4d6f-98a9-8cbb39664d6d"
/>