UI: Uniforming Wattson Displayed Columns
There are 3 levels of Wattson Power Estimates in Perfetto UI: Thread-level, Process-level, and Package-level. There are discrepancies between Package-level displayed columns and the other two. More specifically, package-level is having an extra Total Duration column and
is missing below columns:
1. Idle transitions overhead (estimated mWs)
2. Total energy (estimated mWs)
3. % of total energy

This change is to update package-level columns to align it with the other two, making corresponding UI and SQL changes.

Test: Tested on local Perfetto UI Build
Bug: 444015074