ui: Add warning icons for missing trace events
In case users missed some necessary ftrace_events for Wattson, Perfetto UI will show nothing which confuses the users. There needs a "warning" icon and text telling user which bare minimum ftrace_events he forgot adding in Perfetto Trace Config. It's per track and not for the whole Wattson group because the missing traces might be different per track.

For short-term, this patchset is reusing the "help" icon to display the warning messages. The missing events will be generated by comparing between bare minimum trace events and user's configured events. Once generated, missing events will be passed to each CPU Estimate Track and display in red color: "Perfetto trace configuration is missing below events for Wattson to work : xxx"  Limitation is that user needs to hover over the track and click on the icon to display the message. In long term, there should be a dedicated icon to be implemented.

Test: Local Perfetto UI Test
Bug: 444535474