exp: Modify Columns node (#2923)
This commit introduces a new "Modify Columns" node to the Explore Page
query builder.

* New Functionality: Users can now add a "Modify Columns" node to their
query graph. This node allows them to:
       * Select, reorder, and rename columns from the previous node.
       * Add new columns defined by SQL expressions.
       * Filter the results based on the final set of columns.
* UI Changes: A "Modify Columns" option has been added to the menu when
adding a new node in the query builder UI.
* Serialization: The new node's state (which columns are selected, new
column definitions, filters, etc.) can be saved and loaded, so queries
using this feature can be shared and restored.
* Testing: Unit tests have been added to ensure the serialization and
deserialization of the new node work correctly.