ui: Add self-time (aka self-duration) to slice aggregation (#2946)
This patch adds a new column in the slice aggregation table for 'Self
duration', which is the time a slice was running minus the total time
any of its children were running.

Note: The additional column adds around 25% overhead to how long it
takes to calculate the entire aggregation, which is reasonable
considering the the extra functionality.

<img width="372" height="542" alt="image"
src="https://github.com/user-attachments/assets/44c26334-eba8-4730-944b-068ad6161850"
/>