ui: stabilize macro commands to run on startup (#2908)
If the user defines a macro, as long as all the commands used inside are
allowlisted, there should be no problem with also allowlisting the macro
itself.

This allows for people to define a macro and then make a startup command
out of that macro which I think would be a reasonably common usecase.