ui: update location of internal user script loader (#2906)
This script is being moved due to some automation changes. Use the old
and new versions of the script. We will leave the old one where it is
forever, it just won't be changed so there will be no impact on old
versions of the UI.

Corresponds to internal change cl/806014314. Should not be submitted
until that goes in and successfully releases at least once.