Update suspend.sql (#2968)
In the previous PR, we introduced duration_gap but never used it. This
PR uses that to filter out any overlapping suspends, which tends to
happen when there are data loss.

This is identical to https://github.com/google/perfetto/pull/2883 but I
messed up the branch so now I'm attempting again