ui: Intersect flamegraph slices with area selection (#2977)
Currently when we calculate the set of slices that should appear in the
flamegraph view we simply find the slices that intersect with the area
selection. This patch uses interval intersect to generate a new set of
slices based on the intersection with the area selection window, which
produces a new set of slices whose ts and dur are clamped to the bounds
of the area selection.

This brings flamegraph in line with how the other aggregations work.