perfetto: setup nightly LUCI builds
This CL makes multiple changes to enable nightly LUCI builds to be
possible and useful:
1) It sets up the poller to run nightly to trigger builds
2) It adds new "latest" and "latest-tagged" folders on GCS to have
   a stable place to store builds instead of relying on commit
   SHA
3) For LUCI, it adds new "latest-tagged" to distinguish between
   official releases and just nightly builds (the latest build
   always gets the "latest" tag)
4) Updates all necessary deps (depot_tools, recipe modules etc)

The main reason for doing this is to let people who depend on
Perfetto tools and want to run against those on CI to do so. Also
for our own sake, it lets us catch breakages in Windows bots much
faster which is nice.