ui: add "name or path" option to *ByRegex* track commands
This CL adds a second/third parameter to the public track commands which
filter by regex to indicate whether the filtering should happen on the
track name (the default) or the track path (i.e. a `>` separated list of
track names including the ancestors).

This allows for scoping down e.g. pinning a given thread name even if
that thread appears in lots of processes.

Fixes: https://github.com/google/perfetto/issues/3007