ui: disable prompts during macro execution (if not already disabled)
Was playing around with macros and found that we could end up prompting
the user if there is an optional argument (e.g. with the "nameOrPath"
options for all the track commands) even when inside a macro. This is
unexpected as macros are meant to be self-contained, fully automatic
snippets: any prompts during that time should be blocked.