Add 'base::CopyFileContents' function. (#2971)
We doesn't expect `CopyFileContents` to be a bottleneck, so the
implementation is the same for all platforms.
In the future it may be useful to add a platform-specific
implementations, that uses `sendfile` or even `copy_file_range` APIs.

`CopyFile` have three different ways to fail, so it returns human
readable `base::Status` instead of just `bool` or `ssize_t`.