ui: Remove `traceInfo.cpus`. (#3018)
Currently the core loads a list of CPUs within loadTrace() from the
`cpu` table. This is exposed and used by various plugins in a slightly
strange way.

This patch removes this from loadTrace() and thus removes knowledge
about CPUs from the core entirely, which is good for isolation. CPUs are
similar to threads and processes insofar that the core shouldn't know
anything about them - all cpu related info should be provided by
plugins.

In most of the plugins that use the cpu data to e.g. augment their
tracks with extra data, this can simply be replaced with a join on the
`cpu` table in their original query to look up the tracks, which has
minimal overhead.

There is one case - Wattson - that actually needs the number of cpus
with sched data - i.e. the number of sched tracks. For this I've exposed
this info from the sched plugin and modified Wattson to consume this
information via a plugin dependency on the sched plugin. Wattson is
already very interested in the sched plugin as it adds aggregations to
the sched tracks, so I think an explicit dependency is fine.