ui: Allow tracks to have multiple kind tags (#3020)
Currently tracks can only have one `kind` tag. This is limiting as we
might want to, for example, treat a track as a slice track for the sake
of searching, but treat it like a frames track for the sake of
aggregation.

This patch switches out the singular `kind` tag with a new `kinds` tag
which is an array of strings. Now, tracks can have multiple kinds that
bulk operations can use to select tracks they operate on. This is very
similar to how CSS selectors and classes work.

This solves the concrete problem of allowing frame slice tracks to be
searched using dataset search, while still allowing them to have the
specialist frame aggregation.