stdlib: Fix small "rounding" errors
Remove the removal of doing calculations for time slices less than 1us;
in other words, do calculations for every time slice.

The difference in power calculations are negligible (less than 0.01%
difference) and far below the noise floor, but this helps ensure that
the cumulative sum of certain columns perfectly match up instead being
slightly off, which can be a source of confusion.

The runtime for Wattson calculations are up to 1% slower now, but that
is an acceptable tradeoff for what we are getting in return.

Bug: 435479190
Test: tools/diff_test_trace_processor.py out/linux_clang_release/trace_processor_shell --name-filter '.*wattson*' --print-slowest-tests
Signed-off-by: Samuel Wu <wusamuel@google.com>