tp: readd render stage queue name dimension to tracks
In https://github.com/google/perfetto/pull/2477, we changed render stage
events to only take into account the queue id not the queue name.
Unfortunately, this was not a correct change in practice: turns out
queues use both name *and* id as a dimension.

Readd name as a breakdown for tracks. There's some subtlty about how we
deal with names which are only specified later: we need to preserve the
original dimension as dimensions cannot be changed later.

Fixes: https://github.com/google/perfetto/issues/3030