ui: Fix disappearing slices by explicitly ordering slices
Currently we make the assumption in BaseSliceTrack that slices are
ordered by ts when rendering them, which allows us to cull slices that
are outside of the viewport.

However, there is no explicit order by guarantees in the track query,
so this ordering is not a given, it just happens to be that most
track sources are inherently ordered.

This patch simply adds an `ORDER BY tsQ` clause to the query, forcing
the results to be ordered by timetstamp.

Change-Id: Ieb76f56fe893543701f9d73c5f41394b8f6baf69