ui: Fix disappearing slices by removing an optimization that assumes sorted slices in BaseSliceTrack (#3044)
Currently we make the assumption in BaseSliceTrack that slices are
ordered by `ts`, which allows us to perform a very specific optimization
which avoids rendering anything when the first slice is off the right
hand side of the viewport.

However, slices from the mipmap function are ordered by depth first then
ts, meaning that slices are not necessarily sorted by ts. In practice,
slices at lower depths usually do start before those at higher depths
anyway, so this is usually not an issue.

This patch removes the optimization entirely as it seems pretty dubious
whether it actually helps all that much.

Fixes: https://buganizer.corp.google.com/issues/446503267