ui: Improve HotkeyContext for embeddability (#3028)
Turn it into a proper div in the DOM, and attach key events to this
instead of to the document. This means that the hotkeys are only
triggered when the the context element is focuesed. This makes the UI
more friendly to embedded contexts.

In order to maintain behavior in the normal build of the UI, this
context is automatically focused on creation, which means hotkeys work
without having to focus anything first on ui.perfetto.dev.