Fix bug in PerfettoApiTest#NoFlushFlag/System
Fix a test. This test has always been buggy as it was missing
a sync fence. It was passing by accident just because the
old task runner was slower.
The new task runner makes the bug visible because runs tasks
faster.
This can be repro locally by setting
enable_perfetto_lockfree_taskrunner=true

Bug: 447118803