traced_perf: add event modifiers such as :u and :k (#3016)
It's sometimes useful to configure a perf event to only increment while
in a specific context (userspace vs kernel vs hypervisor). This patch
adds "event modifiers" to both the timebase and the followers to achieve
this, modeled after perftool's
https://man7.org/linux/man-pages/man1/perf-list.1.html#EVENT_MODIFIERS.

The modifiers are recorded in perf_sample_defaults and contribute to the
trace_processor track name as a suffix.

Example data source config:
```
  data_sources {
    config {
      name: "linux.perf"
      perf_event_config {
        timebase {
          frequency: 1000
          counter: SW_CPU_CLOCK
          timestamp_clock: PERF_CLOCK_MONOTONIC
        }
        followers {
          counter: HW_CPU_CYCLES
          modifiers: EVENT_MODIFIER_COUNT_USERSPACE
        }
        followers {
          counter: HW_CPU_CYCLES
          modifiers: EVENT_MODIFIER_COUNT_KERNEL
        }
      }
    }
  }
```

Based on a contribution by: Ali Utku Selen