ui: Create a single entry point for slice tracks based on DST (#3096)
Currently we have two (main) entry points for creating slice tracks,
that have very different parameters and behaviors:
1. `new DatasetSliceTrack()`
2. `createQuerySliceTrack()`

> Note: We have a lot more bespoke factory functions which are based on
DatasetSliceTrack such as createTraceProcessorTrack(), but they are
special purpose dedicated track factories and usually contained within
the plugins that use them. They are not general purpose factories to
create a track from a query.

This patch normalizes the entry points for creating a slice track from a
query.

- Add a default details panel to DatasetSliceTrack if none provided
which simply lists the column names and values in the provided schema
for the selected slice.
- Relax the requirement to pass an `id` column to `DatasetSliceTrack`.
If omitted, an `id` column will be generated using `ROW_NUMBER()`.
- Add two new static factory functions to DatasetSliceTrack:
- `DatasetSliceTrack.createMaterialized()` - Materializes a perfetto
table from the provided dataset - this function is async (returns a
promise) as the table is materialized eagerly.
- `DatasetSliceTrack.create()` - A direct alias for `new
DatasetSliceTrack()`, exists purely for symmetry with
`createMaterialized()`. Soon we will make the constructor private so
`create` must be used instead, but this is out of scope for this patch.
This function is synchronous.
- Replace all calls to `createQuerySliceTrack()` with
`DatasetSliceTrack.createMaterialized()`, modifying the paramters
accordingly. Some of them arguably don't need to materialize a table and
could use `DatasetSliceTrack.create()` instead but that decision is out
of scope for this patch.
- Remove `createQuerySliceTrack()` entirely.

Migration FAQ: 
- Q: With createQuerySliceTrack, I could provide column mappings for ts,
dur, and name. How do I do that with DST?
A: No one was actually using that feature. If you need to map these
columns, do it in the query.
- Q: How do I provide extra columns to be displayed in the details
panel?
A: Add them to the schema. If you don't know or care about the type of
these columns, use `UNKOWN`.