tools: add tool to update a github discussion and update agents
We are experimenting with using Github discussions to store
semi-permanent notes about changes we want to make to Perfetto. To this
end, to let AI easily interact with these discussions, it's nice to have
a tool to take care of these updates (while correctly dealing with JSON
escaping, the correct APIs to use etc).