protozero: add iterator which can deal with both packed/unpacked fields
This CL introduces a new wrapper class for cases where we need to
actually strictly adhere to the proto spec and deal with both packed and
non-packed fields in the same field. While no native Perfetto tools
writes protos like this, it is the case that we cannot control all the
protos we parse.

For example, pprofs have a few repeated fields which are inconsistent in
how their writers write them. To that end, make it possible to parse
both.