ui: Descriptions for TraceProcessorTrack tracks (#3125)
This patch allows descriptors to be attached to tracks added by the
`dev.perfetto.TraceProcessorTrack` plugin.

This plugin bulk adds certain tracks it discovers in the trace, while
adding some opinionated tweaks on top depending on the track type.

This patch:
1) Plumbs through descriptions in the trace if the track has one (e.g.
if it has a `description` arg in its `source_arg_id` arg set).
2) Adds the ability to override this in the track descriptors.
(counter_tracks.ts and slice_tracks.ts).


For example, if someone wanted to add a description to the process
memory tracks, they might add a new description field to this descriptor
in `counter_tracks.ts`.

```ts
{
    type: 'process_memory',
    topLevelGroup: 'PROCESS',
    group: undefined,
    description:
      ({name}) =>
      () =>
        m(
          'div',
          `Some description for track '${name}'`,
          m(Anchor, {href: 'https://example.com'}, 'example link'),
        ),
  },
```