ui: name perf sample tracks after the counter (#2741)
The aim is to better represent what counter is being sampled on. For
example there are android lab traces that record binder callstacks,
using a tracepoint as the subject of the profiling. Current track names
don't reflect that.

Also, make the panel auto-open only if there's exactly one profiled
process.

Bug: 441579549