ui: Retain page when loading a trace
The UI currently always switches to the timeline page after loading a
trace, which makes it impossible to share a link to a specific page.
Now that more and more pages are being added to the UI, some pages are
going to be equally as important as the timeline page, so it no longer
makes sense to switch to hard switch to the timeline page.

This patch makes it so that we attempt to preseve the current page,
only switching to the timeline page if we are currently on the home
page.

Change-Id: I3d7bf49f726adc0237193537dbcae4cbee75b144