ui: Hide a popup when its trigger is no longer visible
Since popups are rendered out-of-tree, if a trigger becomes invisible
due to it or one of its parents having display:none set, then the
popup itself doesn't automatically become invisible.

This causes issues when e.g. changing page. Since the previous page is
kept in the DOM just with display:none, the trigger element is not
removed, just made invisible. However the popup still persists and
is left floating awkwardly anchored to (0, 0) because the trigger
elements DOM rect is now at the origin.

This patch adds a Popper.js modifier which detects if the trigger
element is visible or not with element.checkVisibility(), and applies
display:none to the popup if not.

Change-Id: I5a30bd9bebe5c0ba636ba62e76bed6cac24ee11f