ui: Hide a popup when its trigger is no longer visible (#3146)
Since popups are rendered out-of-tree, if a trigger becomes invisible
due to it or one of its parents having display:none set, then the popup
itself doesn't automatically become invisible.

This causes issues when e.g. changing page. Since the previous page is
kept in the DOM just with `display:none`, a trigger element with an open
popup on that page is not actually removed, just made invisible. However
the popup still persists and is left floating awkwardly anchored to (0,
0).

This patch adds a popper.js modifier which detects if the trigger
element is visible or not with `element.checkVisibility()`, applying
`display:none` to the popup as well.