trace_processor: add support for print_args(arg_set_id) (#3031)
Add a helper function to print the entire set of args at once as a JSON
object,
reusing existing functionality for exporting arg sets to json.

This makes it easier to inspect arg sets at a glance and would enable
simplifying the UI code by removing the logic which constructs arg tree
in
javascript.