ui: return a trace promise from the openTraceFromXyz() APIs
Return promises of the future trace object to callers of the APIs on the
App that open traces so that they may act upon the opened trace.

Also, an application that embeds the Perfetto UI may need to hand over a
stream of trace data obtained in its own way from a non-URL source and
not gathered into a buffer.

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>