stdlib: Add functionality to extrapolate cpu_idle state (#3178)
Extrapolate or infer the cpuidle state based on the swapper thread (aka
idle thread) when cpu_idle trace events are not available. This allows
for a best effort power estimate of AoT traces and other traces that may
not have the cpu_idle event.

Additionally, make corresponding changes to metrics and UI that rely
upon Wattson, so that there is an indication of when the cpuidle state
is based on the real trace event or when it's extrapolated.

Bug: 435479190
Signed-off-by: Samuel Wu <wusamuel@google.com>