metrics: Maintain original protobuf tag order
Protobuf uses the tag order (e.g. = 1, = 2) and not the field name to
identify fields. Revert back to the original tag order, and move the
newly added tag (from https://github.com/google/perfetto/pull/3178)
to be appended to the original tag order.

Some teams use the tag order and others use the field name to identify
key-value pairs, so moving forward, both need to be consistent for
backwards compatability.

Bug: 371634124
Signed-off-by: Samuel Wu <wusamuel@google.com>