tprobes: generic ftrace events: include group name (#3189)
When using self-describing proto messages for generic (= not known to
perfetto at compile-time) ftrace events, we currently emit, per ftrace
event bundle, a mapping from proto_id to a descriptor of the serialised
proto message (see #1257). Currently, this ends up losing information
about the "group" of the event (e.g. "sched" in "sched/sched_switch"). 

This patch adds the group as a separate proto string field. It's not
part of the payload-describing `DescriptorProto` as I haven't found a
subfield that makes sense. Putting the full "sched/sched_switch" into
`DescriptorProto.name` would: no longer make it a legal C identifier
name; require string splitting in trace_processor; and deviate from
traces written by existing code.

Impl-wise: I've added a wrapper struct for the FlatHashMap as I found
the existing code ugly. Trying to use type aliases was a losing battle
due to edge cases with forward declarations and nested types.

Note: there are no current plans for this field in trace_processor, but
it's best to retain this information in the trace for the future.