tprobes: ftrace: enable denser_generic_event_encoding by default (#3190)
Attempt to roll out self-describing protos by changing the default. It's
been tested locally, and the only known edge case is with using generic
events from multiple concurrent tracefs instances. I'm working on
addressing that, but we also need to discover any potential users that
are too attached to the specifics of ftrace packet encoding. Setting the
config field explicitly is a way to get back to the old encoding.