ui: add API for command filtering for embedders
In some applications that embed the Perfetto UI, there are a number of
commands that are not needed. For example, in an application that
manages the opening of traces into instances of the Perfetto UI, there
is no need for the commands that open and close traces. And other
commands may overlap or conflict with other capabilities provided by the
host application, which therefore may wish to exclude those commands.

Add an API to the Registry class to permit an embedding application to
install a filter that quietly prevents registration of unwanted
services. Employ this capability in the CommandManager.

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>