tp: speed up connect flow operator (#3223)
Instead of doing O(n) lookups, cache the data in a HashMap. Ideally we
would have an index directly on the underlying table but that logic is a
bit murky because of the SharedStorage logic so don't use that. Ideally
we would migrate to the index when that is removed/improved.

Bug: https://github.com/google/perfetto/discussions/3220