Reapply "Support clone operation for 'write_into_file' sessions. (#2871)" (#3235)
This PR enables the cloning of `write_into_file` sessions that have
already written their data to disk before the clone is requested.

This PR doesn't change the existing behavior for the Traceur: the clone
operation doesn't trigger the write for the `write_into_file` session.
The tracing session with the very long `file_write_period_ms` will, as
before, write the file only once, at the end of the tracing session.

This PR relies on the `TraceBuffer::CloneReadOnly` to have correct
behavior (see
https://github.com/google/perfetto/commit/c74edc8f40710244d1b42d542af78bdac8dca512),
so the new code
in this PR is guarded by the `buffer_clone_preserve_read_iter` flag,
just like the `CloneReadOnly` changes.

SELinux change was merged: ag/36238847, so we can reapply this
commit now.

Bug: b/382209797

This reverts commit 11a3d5fb3890853111293ec0991ebff03b28a4f6.