ui: Fix pgup/pgdown scrolling in the timeline (#3248)
https://github.com/google/perfetto/pull/3028 added a focusable div
around the entire UI to handle hotkeys more elegantly in embedded
applications. It transpires, however, that when the body is focused, the
browser does some magic such scrolling scrollable containers when
pgup/pgdown is pressed, that it doesn't do when a div is focused, even
if the scrollable element is inside it.

The PR was also causing some other issues such as the focusable div
would lose focus every now and again meaning that hotkeys would randomly
stop working, including when blurring any element programmatically with
`element.blur()`.

This patch adds a `focusable` property to the HotkeyContext component
which, when disabled, makes the HotkeyContext un-focusable and binds the
keydown event handlers to the document instead. This effectively
restores the previous functionality but allows the focusable
functionality to be enabled e.g. for embedded applications.